Renamed AmqpMessageReceiver to MessageReceiver, and refactored accordingly
