+++ src/Microsoft.Azure.ServiceBus/Core/MessageReceiver.cs
@@ -8,15 +8,104 @@ namespace Microsoft.Azure.ServiceBus.Core
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Azure.Amqp;
+    using Azure.Amqp.Encoding;
+    using Azure.Amqp.Framing;
+    using Messaging.Amqp;
+    using Microsoft.Azure.ServiceBus.Amqp;
+    using Primitives;
 
-    public abstract class MessageReceiver : ClientEntity, IMessageReceiver
+    public class MessageReceiver : ClientEntity, IMessageReceiver
     {
+        public static readonly TimeSpan DefaultBatchFlushInterval = TimeSpan.FromMilliseconds(20);
+
+        protected string sessionId;
+        protected DateTime lockedUntilUtc;
+
+        readonly ConcurrentExpiringSet<Guid> requestResponseLockedMessages;
+        readonly string entityPath;
+        readonly bool isSessionReceiver;
         readonly TimeSpan operationTimeout;
         readonly object messageReceivePumpSyncLock;
+        readonly bool ownsConnection;
+
         int prefetchCount;
         long lastPeekedSequenceNumber;
         MessageReceivePump receivePump;
         CancellationTokenSource receivePumpCancellationTokenSource;
+
+        public MessageReceiver(ServiceBusConnectionStringBuilder connectionStringBuilder, RetryPolicy retryPolicy = null)
+            : this(connectionStringBuilder.GetNamespaceConnectionString(), connectionStringBuilder.EntityPath, retryPolicy)
+        {
+        }
+
+        public MessageReceiver(string connectionString, string entityPath, RetryPolicy retryPolicy = null)
+            : this(entityPath, new ServiceBusNamespaceConnection(connectionString), retryPolicy ?? RetryPolicy.Default)
+        {
+            if (string.IsNullOrWhiteSpace(connectionString))
+            {
+                throw Fx.Exception.ArgumentNullOrWhiteSpace(connectionString);
+            }
+            if (string.IsNullOrWhiteSpace(entityPath))
+            {
+                throw Fx.Exception.ArgumentNullOrWhiteSpace(entityPath);
+            }
+
+            this.ownsConnection = true;
+        }
+
+        public MessageReceiver(string entityPath, ServiceBusConnection serviceBusConnection, RetryPolicy retryPolicy)
+            : base(nameof(MessageSender) + StringUtility.GetRandomString(), retryPolicy ?? RetryPolicy.Default)
+        {
+            this.operationTimeout = serviceBusConnection.OperationTimeout;
+            this.messageReceivePumpSyncLock = new object();
+            this.entityPath = entityPath;
+            this.ServiceBusConnection = serviceBusConnection;
+            this.PrefetchCount = prefetchCount;
+            var tokenProvider = TokenProvider.CreateSharedAccessSignatureTokenProvider(serviceBusConnection.SasKeyName, serviceBusConnection.SasKey);
+            this.CbsTokenProvider = new TokenProviderAdapter(tokenProvider, serviceBusConnection.OperationTimeout);
+            this.ReceiveLinkManager = new FaultTolerantAmqpObject<ReceivingAmqpLink>(this.CreateLinkAsync, this.CloseSession);
+            this.RequestResponseLinkManager = new FaultTolerantAmqpObject<RequestResponseAmqpLink>(this.CreateRequestResponseLinkAsync, this.CloseRequestResponseSession);
+            this.requestResponseLockedMessages = new ConcurrentExpiringSet<Guid>();
+        }
+
+        internal MessageReceiver(
+            string entityPath,
+            MessagingEntityType? entityType,
+            ReceiveMode mode,
+            int prefetchCount,
+            ServiceBusConnection serviceBusConnection,
+            ICbsTokenProvider cbsTokenProvider,
+            RetryPolicy retryPolicy)
+            : this(entityPath, entityType, mode, prefetchCount, serviceBusConnection, cbsTokenProvider, null, retryPolicy)
+        {
+        }
+
+        internal MessageReceiver(
+            string entityPath,
+            MessagingEntityType? entityType,
+            ReceiveMode mode,
+            int prefetchCount,
+            ServiceBusConnection serviceBusConnection,
+            ICbsTokenProvider cbsTokenProvider,
+            string sessionId,
+            RetryPolicy retryPolicy,
+            bool isSessionReceiver = false)
+            : base(nameof(MessageReceiver) + StringUtility.GetRandomString(), retryPolicy ?? RetryPolicy.Default)
+        {
+            this.operationTimeout = serviceBusConnection.OperationTimeout;
+            this.entityPath = entityPath;
+            this.EntityType = entityType;
+            this.ServiceBusConnection = serviceBusConnection;
+            this.CbsTokenProvider = cbsTokenProvider;
+            this.sessionId = sessionId;
+            this.isSessionReceiver = isSessionReceiver;
+            this.ReceiveLinkManager = new FaultTolerantAmqpObject<ReceivingAmqpLink>(this.CreateLinkAsync, this.CloseSession);
+            this.RequestResponseLinkManager = new FaultTolerantAmqpObject<RequestResponseAmqpLink>(this.CreateRequestResponseLinkAsync, this.CloseRequestResponseSession);
+            this.requestResponseLockedMessages = new ConcurrentExpiringSet<Guid>();
+            this.PrefetchCount = prefetchCount;
+            this.messageReceivePumpSyncLock = new object();
+        }
 
         protected MessageReceiver(ReceiveMode receiveMode, TimeSpan operationTimeout, RetryPolicy retryPolicy)
             : base(nameof(MessageReceiver) + StringUtility.GetRandomString(), retryPolicy ?? RetryPolicy.Default)
@@ -27,11 +116,13 @@ namespace Microsoft.Azure.ServiceBus.Core
             this.messageReceivePumpSyncLock = new object();
         }
 
-        public abstract string Path { get; }
-
         public ReceiveMode ReceiveMode { get; protected set; }
 
-        public virtual int PrefetchCount
+        /// <summary>
+        /// Get Prefetch Count configured on the Receiver.
+        /// </summary>
+        /// <value>The upper limit of events this receiver will actively receive regardless of whether a receive operation is pending.</value>
+        public int PrefetchCount
         {
             get
             {
@@ -46,10 +137,15 @@ namespace Microsoft.Azure.ServiceBus.Core
                 }
 
                 this.prefetchCount = value;
+                ReceivingAmqpLink link;
+                if (this.ReceiveLinkManager.TryGetOpenedObject(out link))
+                {
+                    link.SetTotalLinkCredit((uint)value, true, true);
+                }
             }
         }
 
-        public virtual long LastPeekedSequenceNumber
+        public long LastPeekedSequenceNumber
         {
             get
             {
@@ -67,6 +163,21 @@ namespace Microsoft.Azure.ServiceBus.Core
             }
         }
 
+        public virtual string Path
+        {
+            get { return this.entityPath; }
+        }
+
+        public DateTime LockedUntilUtc
+        {
+            get { return this.lockedUntilUtc; }
+        }
+
+        public string SessionId
+        {
+            get { return this.sessionId; }
+        }
+
         internal TimeSpan OperationTimeout
         {
             get { return this.operationTimeout; }
@@ -74,19 +185,13 @@ namespace Microsoft.Azure.ServiceBus.Core
 
         internal MessagingEntityType? EntityType { get; set; }
 
-        public override Task OnClosingAsync()
-        {
-            lock (this.messageReceivePumpSyncLock)
-            {
-                if (this.receivePump != null)
-                {
-                    this.receivePumpCancellationTokenSource.Cancel();
-                    this.receivePumpCancellationTokenSource.Dispose();
-                    this.receivePump = null;
-                }
-            }
-            return Task.FromResult(0);
-        }
+        ServiceBusConnection ServiceBusConnection { get; }
+
+        ICbsTokenProvider CbsTokenProvider { get; }
+
+        FaultTolerantAmqpObject<ReceivingAmqpLink> ReceiveLinkManager { get; }
+
+        FaultTolerantAmqpObject<RequestResponseAmqpLink> RequestResponseLinkManager { get; }
 
         /// <summary>
         /// Asynchronously receives a message using the <see cref="MessageReceiver" />.
@@ -388,21 +493,432 @@ namespace Microsoft.Azure.ServiceBus.Core
             this.OnMessageHandlerAsync(messageHandlerOptions, handler).GetAwaiter().GetResult();
         }
 
-        protected abstract Task<IList<Message>> OnReceiveAsync(int maxMessageCount, TimeSpan serverWaitTime);
+        public override async Task OnClosingAsync()
+        {
+            lock (this.messageReceivePumpSyncLock)
+            {
+                if (this.receivePump != null)
+                {
+                    this.receivePumpCancellationTokenSource.Cancel();
+                    this.receivePumpCancellationTokenSource.Dispose();
+                    this.receivePump = null;
+                }
+            }
+            await this.ReceiveLinkManager.CloseAsync().ConfigureAwait(false);
+            await this.RequestResponseLinkManager.CloseAsync().ConfigureAwait(false);
 
-        protected abstract Task<IList<Message>> OnReceiveBySequenceNumberAsync(IEnumerable<long> sequenceNumbers);
+            if (this.ownsConnection)
+            {
+                await this.ServiceBusConnection.CloseAsync().ConfigureAwait(false);
+            }
+        }
 
-        protected abstract Task OnCompleteAsync(IEnumerable<string> lockTokens);
+        internal async Task GetSessionReceiverLinkAsync(TimeSpan serverWaitTime)
+        {
+            TimeoutHelper timeoutHelper = new TimeoutHelper(serverWaitTime, true);
+            ReceivingAmqpLink receivingAmqpLink = await this.ReceiveLinkManager.GetOrCreateAsync(timeoutHelper.RemainingTime()).ConfigureAwait(false);
+            Source source = (Source)receivingAmqpLink.Settings.Source;
+            if (!source.FilterSet.TryGetValue(AmqpClientConstants.SessionFilterName, out this.sessionId))
+            {
+                receivingAmqpLink.Session.SafeClose();
+                throw new ServiceBusException(false, Resources.AmqpFieldSessionId);
+            }
 
-        protected abstract Task OnAbandonAsync(string lockToken);
+            long lockedUntilUtcTicks;
+            this.lockedUntilUtc = receivingAmqpLink.Settings.Properties.TryGetValue(AmqpClientConstants.LockedUntilUtc, out lockedUntilUtcTicks) ? new DateTime(lockedUntilUtcTicks, DateTimeKind.Utc) : DateTime.MinValue;
+        }
 
-        protected abstract Task OnDeferAsync(string lockToken);
+        internal async Task<AmqpResponseMessage> ExecuteRequestResponseAsync(AmqpRequestMessage amqpRequestMessage)
+        {
+            AmqpMessage amqpMessage = amqpRequestMessage.AmqpMessage;
+            TimeoutHelper timeoutHelper = new TimeoutHelper(this.OperationTimeout, true);
+            RequestResponseAmqpLink requestResponseAmqpLink = await this.RequestResponseLinkManager.GetOrCreateAsync(timeoutHelper.RemainingTime()).ConfigureAwait(false);
 
-        protected abstract Task OnDeadLetterAsync(string lockToken);
+            AmqpMessage responseAmqpMessage = await Task.Factory.FromAsync(
+                (c, s) => requestResponseAmqpLink.BeginRequest(amqpMessage, timeoutHelper.RemainingTime(), c, s),
+                (a) => requestResponseAmqpLink.EndRequest(a),
+                this).ConfigureAwait(false);
 
-        protected abstract Task<DateTime> OnRenewLockAsync(string lockToken);
+            AmqpResponseMessage responseMessage = AmqpResponseMessage.CreateResponse(responseAmqpMessage);
+            return responseMessage;
+        }
 
-        protected abstract Task<IList<Message>> OnPeekAsync(long fromSequenceNumber, int messageCount = 1);
+        protected virtual async Task<IList<Message>> OnReceiveAsync(int maxMessageCount, TimeSpan serverWaitTime)
+        {
+            ReceivingAmqpLink receiveLink = null;
+            try
+            {
+                TimeoutHelper timeoutHelper = new TimeoutHelper(serverWaitTime, true);
+                receiveLink = await this.ReceiveLinkManager.GetOrCreateAsync(timeoutHelper.RemainingTime()).ConfigureAwait(false);
+
+                IEnumerable<AmqpMessage> amqpMessages = null;
+                bool hasMessages = await Task.Factory.FromAsync(
+                    (c, s) => receiveLink.BeginReceiveRemoteMessages(maxMessageCount, DefaultBatchFlushInterval, timeoutHelper.RemainingTime(), c, s),
+                    a => receiveLink.EndReceiveMessages(a, out amqpMessages),
+                    this).ConfigureAwait(false);
+
+                if (receiveLink.TerminalException != null)
+                {
+                    throw receiveLink.TerminalException;
+                }
+
+                if (hasMessages && amqpMessages != null)
+                {
+                    IList<Message> brokeredMessages = null;
+                    foreach (var amqpMessage in amqpMessages)
+                    {
+                        if (brokeredMessages == null)
+                        {
+                            brokeredMessages = new List<Message>();
+                        }
+
+                        if (this.ReceiveMode == ReceiveMode.ReceiveAndDelete)
+                        {
+                            receiveLink.DisposeDelivery(amqpMessage, true, AmqpConstants.AcceptedOutcome);
+                        }
+
+                        Message message = AmqpMessageConverter.AmqpMessageToSBMessage(amqpMessage);
+                        brokeredMessages.Add(message);
+                    }
+
+                    return brokeredMessages;
+                }
+
+                return null;
+            }
+            catch (Exception exception)
+            {
+                throw AmqpExceptionHelper.GetClientException(exception, receiveLink?.GetTrackingId());
+            }
+        }
+
+        protected virtual async Task<IList<Message>> OnPeekAsync(long fromSequenceNumber, int messageCount = 1)
+        {
+            try
+            {
+                AmqpRequestMessage requestMessage =
+                    AmqpRequestMessage.CreateRequest(
+                        ManagementConstants.Operations.PeekMessageOperation,
+                        this.OperationTimeout,
+                        null);
+
+                requestMessage.Map[ManagementConstants.Properties.FromSequenceNumber] = fromSequenceNumber;
+                requestMessage.Map[ManagementConstants.Properties.MessageCount] = messageCount;
+
+                if (!string.IsNullOrWhiteSpace(this.sessionId))
+                {
+                    requestMessage.Map[ManagementConstants.Properties.SessionId] = this.sessionId;
+                }
+
+                List<Message> messages = new List<Message>();
+
+                AmqpResponseMessage response = await this.ExecuteRequestResponseAsync(requestMessage).ConfigureAwait(false);
+                if (response.StatusCode == AmqpResponseStatusCode.OK)
+                {
+                    Message message = null;
+                    var messageList = response.GetListValue<AmqpMap>(ManagementConstants.Properties.Messages);
+                    foreach (AmqpMap entry in messageList)
+                    {
+                        var payload = (ArraySegment<byte>)entry[ManagementConstants.Properties.Message];
+                        AmqpMessage amqpMessage =
+                            AmqpMessage.CreateAmqpStreamMessage(new BufferListStream(new[] { payload }), true);
+                        message = AmqpMessageConverter.AmqpMessageToSBMessage(amqpMessage);
+                        messages.Add(message);
+                    }
+
+                    if (message != null)
+                    {
+                        this.LastPeekedSequenceNumber = message.SystemProperties.SequenceNumber;
+                    }
+
+                    return messages;
+                }
+                else if (response.StatusCode == AmqpResponseStatusCode.NoContent ||
+                        (response.StatusCode == AmqpResponseStatusCode.NotFound && Equals(AmqpClientConstants.MessageNotFoundError, response.GetResponseErrorCondition())))
+                {
+                    return messages;
+                }
+                else
+                {
+                    throw response.ToMessagingContractException();
+                }
+            }
+            catch (Exception exception)
+            {
+                throw AmqpExceptionHelper.GetClientException(exception);
+            }
+        }
+
+        protected virtual async Task<IList<Message>> OnReceiveBySequenceNumberAsync(IEnumerable<long> sequenceNumbers)
+        {
+            List<Message> messages = new List<Message>();
+            try
+            {
+                AmqpRequestMessage requestMessage = AmqpRequestMessage.CreateRequest(ManagementConstants.Operations.ReceiveBySequenceNumberOperation, this.OperationTimeout, null);
+                requestMessage.Map[ManagementConstants.Properties.SequenceNumbers] = sequenceNumbers.ToArray();
+                requestMessage.Map[ManagementConstants.Properties.ReceiverSettleMode] = (uint)(this.ReceiveMode == ReceiveMode.ReceiveAndDelete ? 0 : 1);
+
+                AmqpResponseMessage response = await this.ExecuteRequestResponseAsync(requestMessage).ConfigureAwait(false);
+
+                if (response.StatusCode == AmqpResponseStatusCode.OK)
+                {
+                    IEnumerable<AmqpMap> amqpMapList = response.GetListValue<AmqpMap>(ManagementConstants.Properties.Messages);
+                    foreach (AmqpMap entry in amqpMapList)
+                    {
+                        ArraySegment<byte> payload = (ArraySegment<byte>)entry[ManagementConstants.Properties.Message];
+                        AmqpMessage amqpMessage = AmqpMessage.CreateAmqpStreamMessage(new BufferListStream(new[] { payload }), true);
+                        Message message = AmqpMessageConverter.AmqpMessageToSBMessage(amqpMessage);
+                        Guid lockToken;
+                        if (entry.TryGetValue(ManagementConstants.Properties.LockToken, out lockToken))
+                        {
+                            message.SystemProperties.LockTokenGuid = lockToken;
+                            this.requestResponseLockedMessages.AddOrUpdate(lockToken, message.SystemProperties.LockedUntilUtc);
+                        }
+
+                        messages.Add(message);
+                    }
+                }
+                else
+                {
+                    throw response.ToMessagingContractException();
+                }
+            }
+            catch (Exception exception)
+            {
+                throw AmqpExceptionHelper.GetClientException(exception);
+            }
+
+            return messages;
+        }
+
+        protected virtual async Task OnCompleteAsync(IEnumerable<string> lockTokens)
+        {
+            var lockTokenGuids = lockTokens.Select(lt => new Guid(lt));
+            if (lockTokenGuids.Any(lt => this.requestResponseLockedMessages.Contains(lt)))
+            {
+                await this.DisposeMessageRequestResponseAsync(lockTokenGuids, DispositionStatus.Completed).ConfigureAwait(false);
+            }
+            else
+            {
+                await this.DisposeMessagesAsync(lockTokenGuids, AmqpConstants.AcceptedOutcome).ConfigureAwait(false);
+            }
+        }
+
+        protected virtual async Task OnAbandonAsync(string lockToken)
+        {
+            IEnumerable<Guid> lockTokens = new[] { new Guid(lockToken) };
+            if (lockTokens.Any((lt) => this.requestResponseLockedMessages.Contains(lt)))
+            {
+                await this.DisposeMessageRequestResponseAsync(lockTokens, DispositionStatus.Abandoned).ConfigureAwait(false);
+            }
+            else
+            {
+                await this.DisposeMessagesAsync(lockTokens, new Modified()).ConfigureAwait(false);
+            }
+        }
+
+        protected virtual async Task OnDeferAsync(string lockToken)
+        {
+            IEnumerable<Guid> lockTokens = new[] { new Guid(lockToken) };
+            if (lockTokens.Any((lt) => this.requestResponseLockedMessages.Contains(lt)))
+            {
+                await this.DisposeMessageRequestResponseAsync(lockTokens, DispositionStatus.Defered).ConfigureAwait(false);
+            }
+            else
+            {
+                await this.DisposeMessagesAsync(lockTokens, new Modified() { UndeliverableHere = true }).ConfigureAwait(false);
+            }
+        }
+
+        protected virtual async Task OnDeadLetterAsync(string lockToken)
+        {
+            IEnumerable<Guid> lockTokens = new[] { new Guid(lockToken) };
+            if (lockTokens.Any((lt) => this.requestResponseLockedMessages.Contains(lt)))
+            {
+                await this.DisposeMessageRequestResponseAsync(lockTokens, DispositionStatus.Suspended).ConfigureAwait(false);
+            }
+            else
+            {
+                await this.DisposeMessagesAsync(lockTokens, AmqpConstants.RejectedOutcome).ConfigureAwait(false);
+            }
+        }
+
+        protected virtual async Task<DateTime> OnRenewLockAsync(string lockToken)
+        {
+            DateTime lockedUntilUtc = DateTime.MinValue;
+            try
+            {
+                // Create an AmqpRequest Message to renew  lock
+                AmqpRequestMessage requestMessage = AmqpRequestMessage.CreateRequest(ManagementConstants.Operations.RenewLockOperation, this.OperationTimeout, null);
+                requestMessage.Map[ManagementConstants.Properties.LockTokens] = new[] { new Guid(lockToken) };
+
+                AmqpResponseMessage response = await this.ExecuteRequestResponseAsync(requestMessage).ConfigureAwait(false);
+
+                if (response.StatusCode == AmqpResponseStatusCode.OK)
+                {
+                    IEnumerable<DateTime> lockedUntilUtcTimes = response.GetValue<IEnumerable<DateTime>>(ManagementConstants.Properties.Expirations);
+                    lockedUntilUtc = lockedUntilUtcTimes.First();
+                }
+                else
+                {
+                    throw response.ToMessagingContractException();
+                }
+            }
+            catch (Exception exception)
+            {
+                throw AmqpExceptionHelper.GetClientException(exception);
+            }
+
+            return lockedUntilUtc;
+        }
+
+        async Task DisposeMessagesAsync(IEnumerable<Guid> lockTokens, Outcome outcome)
+        {
+            TimeoutHelper timeoutHelper = new TimeoutHelper(this.OperationTimeout, true);
+            IList<ArraySegment<byte>> deliveryTags = this.ConvertLockTokensToDeliveryTags(lockTokens);
+
+            ReceivingAmqpLink receiveLink = null;
+            try
+            {
+                receiveLink = await this.ReceiveLinkManager.GetOrCreateAsync(timeoutHelper.RemainingTime()).ConfigureAwait(false);
+                Task<Outcome>[] disposeMessageTasks = new Task<Outcome>[deliveryTags.Count];
+                int i = 0;
+                foreach (ArraySegment<byte> deliveryTag in deliveryTags)
+                {
+                    disposeMessageTasks[i++] = Task.Factory.FromAsync(
+                        (c, s) => receiveLink.BeginDisposeMessage(deliveryTag, outcome, true, timeoutHelper.RemainingTime(), c, s),
+                        a => receiveLink.EndDisposeMessage(a),
+                        this);
+                }
+
+                Outcome[] outcomes = await Task.WhenAll(disposeMessageTasks).ConfigureAwait(false);
+                Error error = null;
+                foreach (Outcome item in outcomes)
+                {
+                    var disposedOutcome = item.DescriptorCode == Rejected.Code && ((error = ((Rejected)item).Error) != null) ? item : null;
+                    if (disposedOutcome != null)
+                    {
+                        if (error.Condition.Equals(AmqpErrorCode.NotFound))
+                        {
+                            if (this.isSessionReceiver)
+                            {
+                                throw new SessionLockLostException(Resources.SessionLockExpiredOnMessageSession);
+                            }
+                            else
+                            {
+                                throw new MessageLockLostException(Resources.MessageLockLost);
+                            }
+                        }
+
+                        throw AmqpExceptionHelper.ToMessagingContractException(error);
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                if (exception is OperationCanceledException &&
+                    receiveLink != null && receiveLink.State != AmqpObjectState.Opened)
+                {
+                    if (this.isSessionReceiver)
+                    {
+                        throw new SessionLockLostException(Resources.SessionLockExpiredOnMessageSession);
+                    }
+                    else
+                    {
+                        throw new MessageLockLostException(Resources.MessageLockLost);
+                    }
+                }
+
+                throw AmqpExceptionHelper.GetClientException(exception);
+            }
+        }
+
+        async Task DisposeMessageRequestResponseAsync(IEnumerable<Guid> lockTokens, DispositionStatus dispositionStatus)
+        {
+            try
+            {
+                // Create an AmqpRequest Message to update disposition
+                AmqpRequestMessage requestMessage = AmqpRequestMessage.CreateRequest(ManagementConstants.Operations.UpdateDispositionOperation, this.OperationTimeout, null);
+                requestMessage.Map[ManagementConstants.Properties.LockTokens] = lockTokens.ToArray();
+                requestMessage.Map[ManagementConstants.Properties.DispositionStatus] = dispositionStatus.ToString().ToLowerInvariant();
+
+                AmqpResponseMessage amqpResponseMessage = await this.ExecuteRequestResponseAsync(requestMessage).ConfigureAwait(false);
+                if (amqpResponseMessage.StatusCode != AmqpResponseStatusCode.OK)
+                {
+                    throw amqpResponseMessage.ToMessagingContractException();
+                }
+            }
+            catch (Exception exception)
+            {
+                throw AmqpExceptionHelper.GetClientException(exception);
+            }
+        }
+
+        IList<ArraySegment<byte>> ConvertLockTokensToDeliveryTags(IEnumerable<Guid> lockTokens)
+        {
+            return lockTokens.Select(lockToken => new ArraySegment<byte>(lockToken.ToByteArray())).ToList();
+        }
+
+        async Task<ReceivingAmqpLink> CreateLinkAsync(TimeSpan timeout)
+        {
+            FilterSet filterMap = null;
+
+            MessagingEventSource.Log.AmqpReceiveLinkCreateStart(this.ClientId, false, this.EntityType, this.Path);
+
+            if (this.isSessionReceiver)
+            {
+                filterMap = new FilterSet { { AmqpClientConstants.SessionFilterName, this.sessionId } };
+            }
+
+            AmqpLinkSettings linkSettings = new AmqpLinkSettings
+            {
+                Role = true,
+                TotalLinkCredit = (uint)this.PrefetchCount,
+                AutoSendFlow = this.PrefetchCount > 0,
+                Source = new Source { Address = this.Path, FilterSet = filterMap },
+                SettleType = (this.ReceiveMode == ReceiveMode.PeekLock) ? SettleMode.SettleOnDispose : SettleMode.SettleOnSend
+            };
+
+            if (this.EntityType != null)
+            {
+                linkSettings.AddProperty(AmqpClientConstants.EntityTypeName, (int)this.EntityType);
+            }
+
+            linkSettings.AddProperty(AmqpClientConstants.TimeoutName, (uint)timeout.TotalMilliseconds);
+
+            AmqpSendReceiveLinkCreator sendReceiveLinkCreator = new AmqpSendReceiveLinkCreator(this.Path, this.ServiceBusConnection, new[] { ClaimConstants.Listen }, this.CbsTokenProvider, linkSettings);
+            ReceivingAmqpLink receivingAmqpLink = (ReceivingAmqpLink)await sendReceiveLinkCreator.CreateAndOpenAmqpLinkAsync().ConfigureAwait(false);
+
+            MessagingEventSource.Log.AmqpReceiveLinkCreateStop(this.ClientId);
+
+            return receivingAmqpLink;
+        }
+
+        // TODO: Consolidate the link creation paths
+        async Task<RequestResponseAmqpLink> CreateRequestResponseLinkAsync(TimeSpan timeout)
+        {
+            string entityPath = this.Path + '/' + AmqpClientConstants.ManagementAddress;
+
+            MessagingEventSource.Log.AmqpReceiveLinkCreateStart(this.ClientId, true, this.EntityType, entityPath);
+            AmqpLinkSettings linkSettings = new AmqpLinkSettings();
+            linkSettings.AddProperty(AmqpClientConstants.EntityTypeName, AmqpClientConstants.EntityTypeManagement);
+
+            AmqpRequestResponseLinkCreator requestResponseLinkCreator = new AmqpRequestResponseLinkCreator(entityPath, this.ServiceBusConnection, new[] { ClaimConstants.Manage, ClaimConstants.Listen }, this.CbsTokenProvider, linkSettings);
+            RequestResponseAmqpLink requestResponseAmqpLink = (RequestResponseAmqpLink)await requestResponseLinkCreator.CreateAndOpenAmqpLinkAsync().ConfigureAwait(false);
+
+            MessagingEventSource.Log.AmqpReceiveLinkCreateStop(this.ClientId);
+            return requestResponseAmqpLink;
+        }
+
+        void CloseSession(ReceivingAmqpLink link)
+        {
+            link.Session.SafeClose();
+        }
+
+        void CloseRequestResponseSession(RequestResponseAmqpLink requestResponseAmqpLink)
+        {
+            requestResponseAmqpLink.Session.SafeClose();
+        }
 
         static int ValidateLockTokens(IEnumerable<string> lockTokens)
         {
